Refactor /start command handling; extract message sending and typing action to utility functions for improved readability and maintainability
